Define abstract class for SDL-powered `GameHost`s
